// lib/skyapi/node/model/apiV1PendingTxsTransactionOutputs.ts
/**
 * Skycoin REST API.
 * Skycoin is a next-generation cryptocurrency.
 *
 * The version of the OpenAPI document: 0.27.0
 * Contact: contact@skycoin.net
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export class ApiV1PendingTxsTransactionOutputs {
    'uxid'?: string;
    'dst'?: string;
    'coins'?: string;
    'hours'?: number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "uxid",
            "baseName": "uxid",
            "type": "string"
        },
        {
            "name": "dst",
            "baseName": "dst",
            "type": "string"
        },
        {
            "name": "coins",
            "baseName": "coins",
            "type": "string"
        },
        {
            "name": "hours",
            "baseName": "hours",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return ApiV1PendingTxsTransactionOutputs.attributeTypeMap;
    }
}