// lib/skyapi/node/model/blockSchema.ts
/**
 * Skycoin REST API.
 * Skycoin is a next-generation cryptocurrency.
 *
 * OpenAPI spec version: 0.26.0
 * Contact: contact@skycoin.net
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { BlockSchemaBody } from './blockSchemaBody';
import { BlockVerboseSchemaHeader } from './blockVerboseSchemaHeader';

export class BlockSchema {
    'header'?: BlockVerboseSchemaHeader;
    'body'?: BlockSchemaBody;
    'size'?: number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "header",
            "baseName": "header",
            "type": "BlockVerboseSchemaHeader"
        },
        {
            "name": "body",
            "baseName": "body",
            "type": "BlockSchemaBody"
        },
        {
            "name": "size",
            "baseName": "size",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return BlockSchema.attributeTypeMap;
    }
}