// lib/skyapi/node/model/blockSchemaBody.ts
/**
 * Skycoin REST API.
 * Skycoin is a next-generation cryptocurrency.
 *
 * OpenAPI spec version: 0.26.0
 * Contact: contact@skycoin.net
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export class BlockSchemaBody {
    'txns'?: Array<any>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "txns",
            "baseName": "txns",
            "type": "Array<any>"
        }    ];

    static getAttributeTypeMap() {
        return BlockSchemaBody.attributeTypeMap;
    }
}