// lib/skyapi/node/model/blockVerboseSchemaHeader.ts
/**
 * Skycoin REST API.
 * Skycoin is a next-generation cryptocurrency.
 *
 * OpenAPI spec version: 0.26.0
 * Contact: contact@skycoin.net
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export class BlockVerboseSchemaHeader {
    'seq'?: number;
    'timestamp'?: number;
    'fee'?: number;
    'version'?: number;
    'blockHash'?: string;
    'previousBlockHash'?: string;
    'txBodyHash'?: string;
    'uxHash'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "seq",
            "baseName": "seq",
            "type": "number"
        },
        {
            "name": "timestamp",
            "baseName": "timestamp",
            "type": "number"
        },
        {
            "name": "fee",
            "baseName": "fee",
            "type": "number"
        },
        {
            "name": "version",
            "baseName": "version",
            "type": "number"
        },
        {
            "name": "blockHash",
            "baseName": "block_hash",
            "type": "string"
        },
        {
            "name": "previousBlockHash",
            "baseName": "previous_block_hash",
            "type": "string"
        },
        {
            "name": "txBodyHash",
            "baseName": "tx_body_hash",
            "type": "string"
        },
        {
            "name": "uxHash",
            "baseName": "ux_hash",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return BlockVerboseSchemaHeader.attributeTypeMap;
    }
}