// lib/skyapi/node/model/inlineResponse200.ts
/**
 * Skycoin REST API.
 * Skycoin is a next-generation cryptocurrency.
 *
 * OpenAPI spec version: 0.25.1
 * Contact: contact@skycoin.net
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export class InlineResponse200 {
    'hours'?: number;
    'coins'?: number;
    'uxid'?: string;
    'ownerAddress'?: string;
    'spentBlockSeq'?: number;
    'spentTx'?: string;
    'time'?: number;
    'srcBlockSeq'?: number;
    'srcTx'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "hours",
            "baseName": "hours",
            "type": "number"
        },
        {
            "name": "coins",
            "baseName": "coins",
            "type": "number"
        },
        {
            "name": "uxid",
            "baseName": "uxid",
            "type": "string"
        },
        {
            "name": "ownerAddress",
            "baseName": "owner_address",
            "type": "string"
        },
        {
            "name": "spentBlockSeq",
            "baseName": "spent_block_seq",
            "type": "number"
        },
        {
            "name": "spentTx",
            "baseName": "spent_tx",
            "type": "string"
        },
        {
            "name": "time",
            "baseName": "time",
            "type": "number"
        },
        {
            "name": "srcBlockSeq",
            "baseName": "src_block_seq",
            "type": "number"
        },
        {
            "name": "srcTx",
            "baseName": "src_tx",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return InlineResponse200.attributeTypeMap;
    }
}