// lib/skyapi/node/model/inlineResponse2002.ts
/**
 * Skycoin REST API.
 * Skycoin is a next-generation cryptocurrency.
 *
 * OpenAPI spec version: 0.25.1
 * Contact: contact@skycoin.net
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { ApiV1ExplorerAddressInputs } from './apiV1ExplorerAddressInputs';
import { ApiV1ExplorerAddressOutputs } from './apiV1ExplorerAddressOutputs';
import { ApiV1ExplorerAddressStatus } from './apiV1ExplorerAddressStatus';

export class InlineResponse2002 {
    'outputs'?: Array<ApiV1ExplorerAddressOutputs>;
    'innerHash'?: string;
    'inputs'?: Array<ApiV1ExplorerAddressInputs>;
    'fee'?: number;
    'sigs'?: Array<string>;
    'length'?: number;
    'txid'?: string;
    'type'?: number;
    'status'?: ApiV1ExplorerAddressStatus;
    'timestamp'?: number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "outputs",
            "baseName": "outputs",
            "type": "Array<ApiV1ExplorerAddressOutputs>"
        },
        {
            "name": "innerHash",
            "baseName": "inner_hash",
            "type": "string"
        },
        {
            "name": "inputs",
            "baseName": "inputs",
            "type": "Array<ApiV1ExplorerAddressInputs>"
        },
        {
            "name": "fee",
            "baseName": "fee",
            "type": "number"
        },
        {
            "name": "sigs",
            "baseName": "sigs",
            "type": "Array<string>"
        },
        {
            "name": "length",
            "baseName": "length",
            "type": "number"
        },
        {
            "name": "txid",
            "baseName": "txid",
            "type": "string"
        },
        {
            "name": "type",
            "baseName": "type",
            "type": "number"
        },
        {
            "name": "status",
            "baseName": "status",
            "type": "ApiV1ExplorerAddressStatus"
        },
        {
            "name": "timestamp",
            "baseName": "timestamp",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return InlineResponse2002.attributeTypeMap;
    }
}